+++ src/im/gpinyin/include/matrixsearch.h
@@ -26,7 +26,7 @@
 
 namespace ime_pinyin {
 
-static const size_t kMaxRowNum = kMaxSearchSteps;
+static const unsigned kMaxRowNum = kMaxSearchSteps;
 
 typedef struct {
   // MileStoneHandle objects for the system and user dictionaries.
@@ -118,7 +118,7 @@ typedef struct {
   uint16 spl_start[kMaxRowNum];
   char16 chn_str[kMaxRowNum];       // Chinese string.
   uint16 sublma_start[kMaxRowNum];  // Counted in Chinese characters.
-  size_t sublma_num;
+  unsigned sublma_num;
   uint16 length;                    // Counted in Chinese characters.
 } ComposingPhrase, *TComposingPhrase;
 
@@ -138,20 +138,20 @@ class MatrixSearch {
   static const bool kOnlyUserDictPredict = false;
 
   // The maximum buffer to store LmaPsbItems.
-  static const size_t kMaxLmaPsbItems = 1450;
+  static const unsigned kMaxLmaPsbItems = 1450;
 
   // How many rows for each step.
-  static const size_t kMaxNodeARow = 5;
+  static const unsigned kMaxNodeARow = 5;
 
   // The maximum length of the sentence candidates counted in chinese
   // characters
-  static const size_t kMaxSentenceLength = 16;
+  static const unsigned kMaxSentenceLength = 16;
 
   // The size of the matrix node pool.
-  static const size_t kMtrxNdPoolSize = 200;
+  static const unsigned kMtrxNdPoolSize = 200;
 
   // The size of the DMI node pool.
-  static const size_t kDmiPoolSize = 800;
+  static const unsigned kDmiPoolSize = 800;
 
   // Used to indicate whether this object has been initialized.
   bool inited_;
@@ -177,19 +177,19 @@ class MatrixSearch {
   SpellingParser* spl_parser_;
 
   // The maximum allowed length of spelling string (such as a Pinyin string).
-  size_t max_sps_len_;
+  unsigned max_sps_len_;
 
   // The maximum allowed length of a result Chinese string.
-  size_t max_hzs_len_;
+  unsigned max_hzs_len_;
 
   // Pinyin string. Max length: kMaxRowNum - 1
   char pys_[kMaxRowNum];
 
   // The length of the string that has been decoded successfully.
-  size_t pys_decoded_len_;
+  unsigned pys_decoded_len_;
 
   // Shared buffer for multiple purposes.
-  size_t *share_buf_;
+  unsigned *share_buf_;
 
   MatrixNode *mtrx_nd_pool_;
   PoolPosType mtrx_nd_pool_used_;    // How many nodes used in the pool
@@ -201,13 +201,13 @@ class MatrixSearch {
   DictExtPara *dep_;                 // Parameter used to extend DMI nodes.
 
   NPredictItem *npre_items_;         // Used to do prediction
-  size_t npre_items_len_;
+  unsigned npre_items_len_;
 
   // The starting positions and lemma ids for the full sentence candidate.
-  size_t lma_id_num_;
+  unsigned lma_id_num_;
   uint16 lma_start_[kMaxRowNum];     // Counted in spelling ids.
   LemmaIdType lma_id_[kMaxRowNum];
-  size_t fixed_lmas_;
+  unsigned fixed_lmas_;
 
   // If fixed_lmas_ is bigger than i,  Element i is used to indicate whether
   // the i'th lemma id in lma_id_ is the first candidate for that step.
@@ -227,11 +227,11 @@ class MatrixSearch {
 
   // The starting positions and spelling ids for the first full sentence
   // candidate.
-  size_t spl_id_num_;                // Number of splling ids
+  unsigned spl_id_num_;                // Number of splling ids
   uint16 spl_start_[kMaxRowNum];     // Starting positions
   uint16 spl_id_[kMaxRowNum];        // Spelling ids
   // Used to remember the last fixed position, counted in Hanzi.
-  size_t fixed_hzs_;
+  unsigned fixed_hzs_;
 
   // Lemma Items with possibility score, two purposes:
   // 1. In Viterbi decoding, this buffer is used to get all possible candidates
@@ -239,7 +239,7 @@ class MatrixSearch {
   // 2. When the search is done, this buffer is used to get candiates from the
   // first un-fixed step and show them to the user.
   LmaPsbItem lpi_items_[kMaxLmaPsbItems];
-  size_t lpi_total_;
+  unsigned lpi_total_;
 
   // Assign the pointers with NULL. The caller makes sure that all pointers are
   // not valid before calling it. This function only will be called in the
@@ -262,11 +262,11 @@ class MatrixSearch {
   // The DMI nodes will be kept.
   //
   // Note: this function should not destroy content of pys_.
-  bool reset_search(size_t ch_pos, bool clear_fixed_this_step,
+  bool reset_search(unsigned ch_pos, bool clear_fixed_this_step,
                     bool clear_dmi_this_step, bool clear_mtrx_this_step);
 
   // Delete a part of the content in pys_.
-  void del_in_pys(size_t start, size_t len);
+  void del_in_pys(unsigned start, unsigned len);
 
   // Delete a spelling id and its corresponding Chinese character, and merge
   // the fixed lemmas into the composing phrase.
@@ -274,7 +274,7 @@ class MatrixSearch {
   // This function will update the lemma and spelling segmentation information.
   // The caller guarantees that fixed_lmas_ > 0 and del_spl_pos is within
   // the fixed lemmas.
-  void merge_fixed_lmas(size_t del_spl_pos);
+  void merge_fixed_lmas(unsigned del_spl_pos);
 
   // Get spelling start posistions and ids. The result will be stored in
   // spl_id_num_, spl_start_[], spl_id_[].
@@ -286,8 +286,8 @@ class MatrixSearch {
   // If pfullsent is not NULL, means the full sentence candidate may be the
   // same with the coming lemma string, if so, remove that lemma.
   // The result is sorted in descendant order by the frequency score.
-  size_t get_lpis(const uint16* splid_str, size_t splid_str_len,
-                  LmaPsbItem* lma_buf, size_t max_lma_buf,
+  unsigned get_lpis(const uint16* splid_str, unsigned splid_str_len,
+                  LmaPsbItem* lma_buf, unsigned max_lma_buf,
                   const char16 *pfullsent, bool sort_by_psb);
 
   uint16 get_lemma_str(LemmaIdType id_lemma, char16 *str_buf, uint16 str_max);
@@ -315,23 +315,23 @@ class MatrixSearch {
   // calling this function if necessary.
   //
   // The caller should guarantees that NULL != dep.
-  size_t extend_dmi(DictExtPara *dep, DictMatchInfo *dmi_s);
+  unsigned extend_dmi(DictExtPara *dep, DictMatchInfo *dmi_s);
 
   // Extend dmi for the composing phrase.
-  size_t extend_dmi_c(DictExtPara *dep, DictMatchInfo *dmi_s);
+  unsigned extend_dmi_c(DictExtPara *dep, DictMatchInfo *dmi_s);
 
   // Extend a MatrixNode with the give LmaPsbItem list.
   // res_row is the destination row number.
   // This function does not change mtrx_nd_pool_used_. Please change it after
   // calling this function if necessary.
   // return 0 always.
-  size_t extend_mtrx_nd(MatrixNode *mtrx_nd, LmaPsbItem lpi_items[],
-                        size_t lpi_num, PoolPosType dmi_fr, size_t res_row);
+  unsigned extend_mtrx_nd(MatrixNode *mtrx_nd, LmaPsbItem lpi_items[],
+                        unsigned lpi_num, PoolPosType dmi_fr, unsigned res_row);
 
 
   // Try to find a dmi node at step_to position, and the found dmi node should
   // match the given spelling id strings.
-  PoolPosType match_dmi(size_t step_to, uint16 spl_ids[], uint16 spl_id_num);
+  PoolPosType match_dmi(unsigned step_to, uint16 spl_ids[], uint16 spl_id_num);
 
   bool add_char(char ch);
   bool prepare_add_char(char ch);
@@ -352,9 +352,9 @@ class MatrixSearch {
                 bool splid_end_split, unsigned char splstr_len,
                 unsigned char all_full_id);
 
-  size_t inner_predict(const char16 fixed_scis_ids[], uint16 scis_num,
+  unsigned inner_predict(const char16 fixed_scis_ids[], uint16 scis_num,
                        char16 predict_buf[][kMaxPredictSize + 1],
-                       size_t buf_len);
+                       unsigned buf_len);
 
   // Add the first candidate to the user dictionary.
   bool try_add_cand0_to_userdict();
@@ -380,7 +380,7 @@ class MatrixSearch {
   bool init_fd(int sys_fd, long start_offset, long length,
                const char *fn_usr_dict);
 
-  void set_max_lens(size_t max_sps_len, size_t max_hzs_len);
+  void set_max_lens(unsigned max_sps_len, unsigned max_hzs_len);
 
   void close();
 
@@ -396,7 +396,7 @@ class MatrixSearch {
 
   // Search a Pinyin string.
   // Return value is the position successfully parsed.
-  size_t search(const char *py, size_t py_len);
+  unsigned search(const char *py, unsigned py_len);
 
   // Used to delete something in the Pinyin string kept by the engine, and do
   // a re-search.
@@ -411,45 +411,45 @@ class MatrixSearch {
   // If is_pos_in_splid is false, and pos-th character is in the range for the
   // fixed lemmas or composing string, this function will do nothing and just
   // return the result of the previous search.
-  size_t delsearch(size_t pos, bool is_pos_in_splid,
+  unsigned delsearch(unsigned pos, bool is_pos_in_splid,
                    bool clear_fixed_this_step);
 
   // Get the number of candiates, called after search().
-  size_t get_candidate_num();
+  unsigned get_candidate_num();
 
   // Get the Pinyin string stored by the engine.
   // *decoded_len returns the length of the successfully decoded string.
-  const char* get_pystr(size_t *decoded_len);
+  const char* get_pystr(unsigned *decoded_len);
 
   // Get the spelling boundaries for the first sentence candidate.
   // Number of spellings will be returned. The number of valid elements in
   // spl_start is one more than the return value because the last one is used
   // to indicate the beginning of the next un-input speling.
   // For a Pinyin "women", the returned value is 2, spl_start is [0, 2, 5] .
-  size_t get_spl_start(const uint16 *&spl_start);
+  unsigned get_spl_start(const uint16 *&spl_start);
 
   // Get one candiate string. If full sentence candidate is available, it will
   // be the first one.
-  char16* get_candidate(size_t cand_id, char16 *cand_str, size_t max_len);
+  char16* get_candidate(unsigned cand_id, char16 *cand_str, unsigned max_len);
 
   // Get the first candiate, which is a "full sentence".
   // retstr_len is not NULL, it will be used to return the string length.
   // If only_unfixed is true, only unfixed part will be fetched.
-  char16* get_candidate0(char16* cand_str, size_t max_len,
+  char16* get_candidate0(char16* cand_str, unsigned max_len,
                          uint16 *retstr_len, bool only_unfixed);
 
   // Choose a candidate. The decoder will do a search after the fixed position.
-  size_t choose(size_t cand_id);
+  unsigned choose(unsigned cand_id);
 
   // Cancel the last choosing operation, and return the new number of choices.
-  size_t cancel_last_choice();
+  unsigned cancel_last_choice();
 
   // Get the length of fixed Hanzis.
-  size_t get_fixedlen();
+  unsigned get_fixedlen();
 
-  size_t get_predicts(const char16 fixed_buf[],
+  unsigned get_predicts(const char16 fixed_buf[],
                       char16 predict_buf[][kMaxPredictSize + 1],
-                      size_t buf_len);
+                      unsigned buf_len);
 };
 }
